Implement Minimum Active Threshold for False Motion Reduction
Implement a new user-configurable threshold, Minimum Active Threshold, used for the False Motion Reduction zone type.

Previously, all motion detectors in the zone were required to go active within the activation window, in order for the zone to be considered active.

With this change, Minimum Active Threshold is a user-configurable value that defines the minimum number of motion detectors that need to go active in the zone within the activation window, in order for the zone to be considered active. The default is 2 motion detectors.

For example, if a zone has 4 motion detectors, and the Minimum Active Threshold is set to 3, once any 3 of the motion detectors in the zone are active, the zone will be considered active. If the Minimum Active Threshold is equal to the number of motion detectors in the zone, then the behavior is the same as it was before this change (all motion detectors must be active for the zone to be considered active).